removed, duplicate files, refactored code..
